# src-main/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_material_api_v1_courses__course_id__materials_uploads_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_material_api_v1_courses__course_id__materials_uploads_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LearningMaterialLinkCreate": {
        "additionalProperties": false,
        "properties": {
          "module_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Id"
          },
          "source_url": {
            "format": "uri",
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "source_url"
        ],
        "title": "LearningMaterialLinkCreate",
        "type": "object"
      },
      "LearningMaterialRead": {
        "additionalProperties": false,
        "properties": {
          "content_hash": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Content Hash",
            "type": "string"
          },
          "course_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Course Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "extracted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted At"
          },
          "extraction_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Error"
          },
          "failure_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Stage"
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "indexed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexed At"
          },
          "indexing_status": {
            "$ref": "#/components/schemas/MaterialIndexStatus",
            "default": "pending"
          },
          "mime_type": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Mime Type",
            "type": "string"
          },
          "module_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Id"
          },
          "original_filename": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Filename"
          },
          "processing_revision": {
            "default": 0,
            "minimum": 0.0,
            "title": "Processing Revision",
            "type": "integer"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "storage_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Key"
          }
        },
        "required": [
          "course_id",
          "mime_type",
          "content_hash",
          "id",
          "created_at"
        ],
        "title": "LearningMaterialRead",
        "type": "object"
      },
...
      "MaterialIndexStatus": {
        "enum": [
          "pending",
          "processing",
          "extracted",
          "indexed",
          "failed"
        ],
        "title": "MaterialIndexStatus",
        "type": "string"
      },
...
      "MaterialProcessingRead": {
        "additionalProperties": false,
        "properties": {
          "chunk_count": {
            "minimum": 0.0,
            "title": "Chunk Count",
            "type": "integer"
          },
          "indexed_chunk_count": {
            "minimum": 0.0,
            "title": "Indexed Chunk Count",
            "type": "integer"
          },
          "material": {
            "$ref": "#/components/schemas/LearningMaterialRead"
          },
          "processing_revision": {
            "minimum": 0.0,
            "title": "Processing Revision",
            "type": "integer"
          }
        },
        "required": [
          "material",
          "chunk_count",
          "indexed_chunk_count",
          "processing_revision"
        ],
        "title": "MaterialProcessingRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "QuantumLearn API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses/{course_id}/materials": {
      "get": {
        "operationId": "list_materials_api_v1_courses__course_id__materials_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "module_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module Id"
            }
          },
          {
            "in": "query",
            "name": "indexing_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MaterialIndexStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Indexing Status"
            }
          },
          {
            "in": "cookie",
            "name": "quantumlearn_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quantumlearn Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LearningMaterialRead"
                  },
                  "title": "Response List Materials Api V1 Courses  Course Id  Materials Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Materials",
        "tags": [
          "learning materials"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/materials/links": {
      "post": {
        "operationId": "register_linked_material_api_v1_courses__course_id__materials_links_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "quantumlearn_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quantumlearn Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearningMaterialLinkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningMaterialRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Linked Material",
        "tags": [
          "learning materials"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/materials/uploads": {
      "post": {
        "operationId": "upload_material_api_v1_courses__course_id__materials_uploads_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "module_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module Id"
            }
          },
          {
            "in": "cookie",
            "name": "quantumlearn_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quantumlearn Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_material_api_v1_courses__course_id__materials_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningMaterialRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Material",
        "tags": [
          "learning materials"
        ]
      }
    },
    "/api/v1/courses/{course_id}/materials/{material_id}": {
      "delete": {
        "operationId": "delete_material_api_v1_courses__course_id__materials__material_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "title": "Material Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "quantumlearn_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quantumlearn Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Material",
        "tags": [
          "learning materials"
        ]
      },
      "get": {
        "operationId": "read_material_api_v1_courses__course_id__materials__material_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "title": "Material Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "quantumlearn_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quantumlearn Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningMaterialRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Material",
        "tags": [
          "learning materials"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/materials/{material_id}/process": {
      "post": {
        "operationId": "process_material_api_v1_courses__course_id__materials__material_id__process_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "title": "Material Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "title": "Force",
              "type": "boolean"
            }
          },
          {
            "in": "cookie",
            "name": "quantumlearn_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quantumlearn Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialProcessingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Process Material",
        "tags": [
          "learning materials"
        ]
      }
    },
...
  }
}